requirements.txt
streamlit==1.31.0
chromadb==0.4.22
openai==1.12.0
pymupdf==1.23.0
pandas==2.2.0
python-dotenv==1.0.0
